Upgrade Template to 0.4.0-b400

+++ SakhaTyla.Core/Requests/Files/DeleteFileHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
@@ -34,17 +35,21 @@ namespace SakhaTyla.Core.Requests.Files
             var file = await _fileRepository.GetEntities()
                 .Include(e => e.Group)
                 .Where(e => e.Id == request.Id)
-                .FirstOrDefaultAsync();
+                .FirstOrDefaultAsync(cancellationToken);
             if (file == null)
             {
                 throw new ServiceException(_localizer["{0} {1} not found", _localizer["File"], request.Id]);
             }
             if (file.Group.Type == Enums.FileGroupType.Storage)
             {
+                if (file.Url == null)
+                {
+                    throw new Exception($"File {file.Id} Url is empty");
+                }
                 await _fileStorage.DeleteFileAsync(file.Url);
             }
             _fileRepository.Delete(file);
-            await _unitOfWork.CommitAsync();
+            await _unitOfWork.CommitAsync(cancellationToken);
             return Unit.Value;
         }
